Add SevenZipcmdline decompression, Fix autoupdate, add utility method decompressfile

+++ Knossos.NET/Classes/SevenZipConsoleWrapper.cs
@@ -21,12 +21,12 @@ namespace Knossos.NET
         private bool disposed = false;
         public string versionString = string.Empty;
         private string pathToConsoleExecutable;
-        private Action<string>? progressCallback;
+        private Action<int>? progressCallback;
         private Process? process;
         private bool completedSuccessfully = false;
         private CancellationTokenSource? cancelSource;
 
-        public SevenZipConsoleWrapper(Action<string>? progressCallback = null, CancellationTokenSource? cancelSource = null) 
+        public SevenZipConsoleWrapper(Action<int>? progressCallback = null, CancellationTokenSource? cancelSource = null) 
         {
             this.pathToConsoleExecutable = UnpackExec();
             this.progressCallback = progressCallback;
@@ -40,6 +40,52 @@ namespace Knossos.NET
             {
                 Log.Add(Log.LogSeverity.Error, "SevenZipConsoleWrapper.Constructor", "File does not exist: " + pathToConsoleExecutable);
             }
+        }
+
+        /// <summary>
+        /// Decompress a .zip, .7z or .tar.gz file to a folder using the 7zip cmdline tool
+        /// </summary>
+        /// <param name="sourceFile"></param>
+        /// <param name="destFolder"></param>
+        /// <param name="extractFullPath"></param>
+        /// <returns>true if the descompression was successfull, false otherwise</returns>
+        /// <exception cref="ObjectDisposedException"></exception>
+        public async Task<bool> DecompressFile(string sourceFile, string destFolder, bool extractFullPath = true)
+        {
+            if (disposed)
+                throw new ObjectDisposedException("This object was already disposed.");
+
+            bool isTarGz = false;
+
+            if (sourceFile.EndsWith(".tar.gz", StringComparison.OrdinalIgnoreCase))
+                isTarGz = true;
+
+            string cmdline = "\"" + sourceFile + "\" -aoa -bsp1 -y";
+
+            if (extractFullPath)
+                cmdline = "x " + cmdline;
+            else
+                cmdline = "e " + cmdline;
+
+            var result = await Run(cmdline, destFolder);
+
+            if (isTarGz && result)
+            {
+                sourceFile = Path.Combine(destFolder, Path.GetFileName(sourceFile).Replace(".tar.gz", ".tar"));
+                if (extractFullPath)
+                    cmdline = "x ";
+                else
+                    cmdline = "e ";
+                cmdline += "\"" + sourceFile + "\" -aoa -bsp1 -y";
+                result = await Run(cmdline, destFolder);
+                try
+                {
+                    File.Delete(sourceFile);
+                }
+                catch { }
+            }
+
+            return result;
         }
 
         /// <summary>
@@ -57,7 +103,7 @@ namespace Knossos.NET
             string cmdline = "a -t7z -m0=lzma2 -md=64M -mx=9 -ms=on -bsp1 -y ";
             if (doNotStoreTimestamp)
                 cmdline += "-mtm- ";
-            return await Run(cmdline + destFile, sourceFolder);
+            return await Run(cmdline + "\"" + destFile + "\"", sourceFolder);
         }
 
         /// <summary>
@@ -81,7 +127,7 @@ namespace Knossos.NET
                 cmdline = "a -mx=8 -bsp1 -y ";
                 if (doNotStoreTimestamp)
                     cmdline += "-mtm- ";
-                return await Run(cmdline + destFile + ".tar.gz" + " " + destFile + ".tar", sourceFolder);
+                return await Run(cmdline + "\"" + destFile + ".tar.gz\"" + " " + destFile + ".tar", sourceFolder);
             }
             return r;
         }
@@ -123,7 +169,10 @@ namespace Knossos.NET
         {
             try
             {
-                process?.Kill();
+                if (process != null && process.ExitCode != 0)
+                {
+                    process.Kill();
+                }
             }
             catch (Exception ex)
             {
@@ -275,11 +324,12 @@ namespace Knossos.NET
                         {
                             percentage = match.Groups[1].Value;
                         }
-                        progressCallback?.Invoke(percentage);
+                        progressCallback?.Invoke(int.Parse(percentage));
                     }  
                 }catch { }
                 if(e.Data.Contains("Everything is Ok"))
                 {
+                    progressCallback?.Invoke(100);
                     completedSuccessfully = true;
                 }
             }
@@ -308,7 +358,10 @@ namespace Knossos.NET
             {
                 try
                 {
-                    process.Kill();
+                    if (process.ExitCode != 0)
+                    {
+                        process.Kill();
+                    }
                 }
                 catch { }
                 process.Dispose();
